fix: Remove emotes v2, smart items and worlds for ens owners FFs (#3017)
* fix: Remove emotes v2, smart items and worlds for ens owners FFs

* fix: Use new WorldsListPage
