+++ src/components/Modals/DeployModal/DeployModal.types.ts
@@ -10,7 +10,6 @@ export type Props = ModalProps & {
   currentPoolGroup: PoolGroup | null
   project: Project | null
   scene: Scene | null
-  isWorldsForEnsOwnersEnabled: boolean
 }
 
 export type State = {
@@ -25,7 +24,7 @@ export type Step = {
 }
 
 export type OwnProps = Pick<Props, 'metadata'>
-export type MapStateProps = Pick<Props, 'deployment' | 'currentPoolGroup' | 'project' | 'scene' | 'isWorldsForEnsOwnersEnabled'>
+export type MapStateProps = Pick<Props, 'deployment' | 'currentPoolGroup' | 'project' | 'scene'>
 
 export enum DeployModalView {
   NONE = 'NONE',
