+++ src/components/Modals/DeployModal/DeployToWorld/DeployToWorld.container.ts
@@ -2,7 +2,7 @@ import { connect } from 'react-redux'
 import { push, replace } from 'connected-react-router'
 import { RootState } from 'modules/common/types'
 import { getCurrentProject } from 'modules/project/selectors'
-import { getENSByWallet } from 'modules/ens/selectors'
+import { getENSByWallet, getExternalNamesForConnectedWallet } from 'modules/ens/selectors'
 import { deployToWorldRequest } from 'modules/deployment/actions'
 import { getCurrentMetrics } from 'modules/scene/selectors'
 import { recordMediaRequest } from 'modules/media/actions'
@@ -12,15 +12,18 @@ import { MapDispatch, MapDispatchProps, MapStateProps } from './DeployToWorld.ty
 
 import DeployToWorld from './DeployToWorld'
 
-const mapState = (state: RootState): MapStateProps => ({
-  ensList: getENSByWallet(state),
-  project: getCurrentProject(state) as Project,
-  metrics: getCurrentMetrics(state),
-  deployments: getDeploymentsByWorlds(state),
-  deploymentProgress: getUploadProgress(state),
-  error: getError(state),
-  isLoading: isLoading(state)
-})
+const mapState = (state: RootState): MapStateProps => {
+  return {
+    ensList: getENSByWallet(state),
+    externalNames: getExternalNamesForConnectedWallet(state),
+    project: getCurrentProject(state) as Project,
+    metrics: getCurrentMetrics(state),
+    deployments: getDeploymentsByWorlds(state),
+    deploymentProgress: getUploadProgress(state),
+    error: getError(state),
+    isLoading: isLoading(state)
+  }
+}
 
 const mapDispatch = (dispatch: MapDispatch): MapDispatchProps => ({
   onPublish: (projectId: string, name: string) => {
