+++ src/components/ScenesPage/ScenesPage.types.ts
@@ -20,7 +20,6 @@ export type Props = DefaultProps & {
   sortBy: SortBy
   totalPages: number
   poolList: Pool[] | null
-  isWorldsForENSOwnersEnabled: boolean
   onOpenModal: typeof openModal
   onPageChange: (options: PaginationOptions) => void
   onLoadFromScenePool: typeof loadPoolsRequest
@@ -28,7 +27,7 @@ export type Props = DefaultProps & {
 
 export type MapStateProps = Pick<
   Props,
-  'projects' | 'isFetching' | 'isLoggingIn' | 'page' | 'sortBy' | 'totalPages' | 'didCreate' | 'poolList' | 'isWorldsForENSOwnersEnabled'
+  'projects' | 'isFetching' | 'isLoggingIn' | 'page' | 'sortBy' | 'totalPages' | 'didCreate' | 'poolList'
 >
 export type MapDispatchProps = Pick<Props, 'onOpenModal' | 'onPageChange' | 'onLoadFromScenePool'>
 export type MapDispatch = Dispatch<CallHistoryMethodAction | OpenModalAction | LoadPoolsRequestAction>
