--- src/components/WorldListPage_WorldsForEnsOwnersFeature/WorldListPage.css
@@ -1,191 +1,0 @@
-.WorldListPage .filters {
-  margin-bottom: 16px;
-}
-
-.WorldListPage .filters .ui.header.items-count,
-.WorldListPage .ui.table > thead > tr th {
-  color: var(--text);
-}
-
-.WorldListPage .TableRow {
-  cursor: default;
-}
-
-.WorldListPage .empty-names-container {
-  display: flex;
-  flex-direction: column;
-  text-align: center;
-}
-
-.WorldListPage .empty-names-container .empty-icon {
-  margin-bottom: 30px;
-  height: 87px;
-  width: 87px;
-}
-
-.WorldListPage .empty-names-container .empty-icon.dcl-icon {
-  background-image: url(./images/dcl-empty-state-icon.svg);
-}
-
-.WorldListPage .empty-names-container .empty-icon.ens-icon {
-  background-image: url(./images/ens-empty-state-icon.svg);
-}
-
-.WorldListPage .empty-names-container .empty-title {
-  font-size: 24px;
-  font-weight: 600;
-  line-height: 29px;
-  color: var(--text);
-}
-
-.WorldListPage .empty-names-container .empty-description {
-  font-size: 14px;
-  font-weight: 500;
-  line-height: 24px;
-  color: var(--text);
-  max-width: 612px;
-}
-
-.WorldListPage .empty-names-container .empty-action {
-  margin-top: 40px;
-  width: 372px;
-}
-
-.WorldListPage .Icon.right-round-arrow {
-  margin-bottom: -2px;
-}
-
-.WorldListPage .TableRow .dcl.column.subdomain-wrapper {
-  display: flex;
-  flex: none;
-  flex-direction: row;
-}
-
-.WorldListPage .TableRow .world-url {
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-}
-
-.WorldListPage .TableRow .world-url .right {
-  display: flex;
-  align-items: center;
-  margin-right: 20px;
-}
-
-.WorldListPage .TableRow .world-url .right .link {
-  color: var(--text);
-}
-
-.WorldListPage .TableRow .empty-url {
-  color: #ffffffcc;
-  font-style: italic;
-}
-
-.WorldListPage .TableRow .publish-scene {
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.WorldListPage .TableRow .publish-scene span {
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  overflow: hidden;
-  max-width: 100px;
-}
-
-.WorldListPage .TableRow .publish-scene button {
-  margin-right: 20px;
-  width: 165px;
-}
-
-.WorldListPage .TableRow .world-status.active {
-  color: #34ce76;
-}
-
-.WorldListPage .TableRow .world-status.inactive {
-  color: #cfcdd4;
-}
-
-.WorldListPage .TableRow .world-status.warning {
-  color: var(--danger);
-}
-
-.WorldListPage .TableRow .world-status.blocked {
-  color: var(--error);
-}
-
-.WorldListPage .ui.pagination {
-  width: 100%;
-  display: flex;
-  justify-content: center;
-}
-
-.WorldListPage .ui.pagination .item[type='nextItem'],
-.WorldListPage .ui.pagination .item[type='prevItem'] {
-  position: absolute;
-  background-color: var(--secondary);
-}
-
-.WorldListPage .ui.pagination .item[type='prevItem'] {
-  left: 0px;
-}
-
-.WorldListPage .ui.pagination .item[type='nextItem'] {
-  right: 0px;
-}
-
-.WorldListPage .pagination-section {
-  width: 100%;
-}
-
-.WorldListPage .ui.sub.header.items-count {
-  flex: 1 0 auto;
-  text-transform: uppercase;
-  font-size: 13px;
-  line-height: 18px;
-  font-weight: 500;
-}
-
-.WorldListPage .actions {
-  display: flex;
-  flex: 1 0 auto;
-  align-items: center;
-  justify-content: flex-end;
-  font-size: 13px;
-  line-height: 18px;
-  text-transform: uppercase;
-}
-
-.WorldListPage .actions .ui.button.basic {
-  width: 16px;
-  height: 16px;
-  box-sizing: content-box;
-  padding: 5px 10px;
-  border-radius: 4px;
-  border: solid 1px var(--gray);
-}
-
-.WorldListPage .claim-name .Icon {
-  display: block;
-}
-
-.WorldListPage .ui.dropdown {
-  margin-right: 20px;
-}
-
-.WorldListPage .worlds-storage {
-  margin-bottom: 24px;
-}
-
-.WorldListPage .insufficient-storage-message {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-  background-color: var(--dark-two);
-  border-radius: 10px;
-  padding: 22px;
-  margin-bottom: 24px;
-}
