+++ src/routing/Routes.container.ts
@@ -1,13 +1,12 @@
 import { connect } from 'react-redux'
 import { withRouter } from 'react-router'
 import { RootState } from 'modules/common/types'
-import { getIsMaintenanceEnabled, getIsWorldsForEnsOwnersEnabled } from 'modules/features/selectors'
+import { getIsMaintenanceEnabled } from 'modules/features/selectors'
 import { MapStateProps } from './Routes.types'
 import Routes from './Routes'
 
 const mapState = (state: RootState): MapStateProps => ({
-  inMaintenance: getIsMaintenanceEnabled(state),
-  isWorldsForEnsOwnersEnabled: getIsWorldsForEnsOwnersEnabled(state)
+  inMaintenance: getIsMaintenanceEnabled(state)
 })
 
 const mapDispatch = (_: any) => ({})
