+++ src/routing/Routes.types.ts
@@ -2,10 +2,9 @@ import { RouteComponentProps } from 'react-router'
 
 export type Props = RouteComponentProps & {
   inMaintenance: boolean
-  isWorldsForEnsOwnersEnabled: boolean
 }
 
-export type MapStateProps = Pick<Props, 'inMaintenance' | 'isWorldsForEnsOwnersEnabled'>
+export type MapStateProps = Pick<Props, 'inMaintenance'>
 
 export type State = {
   hasError: boolean
